Conflicts:
	RayTra/Shading.cpp

+++ RayTra/Ray.cpp
@@ -32,6 +32,7 @@ unsigned long Ray::count = 0;
 
 Ray::Ray(Vector3d e, Vector3d d, std::vector<double> r, std::vector<Vector3d> a, int t):
 eye(e), dir(d), ref(r), alpha(a), type(t) {
+//#pragma omp atomic
 	Ray::count++;
 	inv = dir.cwiseInverse();
 	reSign();
@@ -41,7 +42,7 @@ Ray::Ray() {
 }
 
 void Ray::reSign() {
-	sign = (inv.array() < 0);
+	sign = (inv.array() < 0).cast<int>();
 }
 
 Ray::~Ray() {
